+++ lib/karafka/instrumentation/logger.rb
@@ -36,14 +36,11 @@ module Karafka
           .to(STDOUT, file)
       end
 
-      # Makes sure the log directory exists
+      # Makes sure the log directory exists as long as we can write to it
       def ensure_dir_exists
-        dir = File.dirname(log_path)
-        FileUtils.mkdir_p(dir) if writable_dir?
-      end
-      
-      def writable_dir?
-        File.world_writable?(dir) && !Dir.exist?(dir)
+        FileUtils.mkdir_p(File.dirname(log_path))
+      rescue Errno::EACCES
+        nil
       end
 
       # @return [Pathname] Path to a file to which we should log