+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Karafka framework changelog
 
+## Unreleased
+- #545 - Makes sure the log directory exists when is possible (robertomiranda)
+
 ## 1.3.0 (2019-09-09)
 - Drop support for Ruby 2.4
 - YARD docs tags cleanup